backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
onnxruntime
manga-ocr
Pillow
pydantic
pydantic-settings
aiofiles
python-dotenv
pyinstaller